Removing checkpoints from mantle timestepping
