+++ docs/conf.py
@@ -13,8 +13,8 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('../../pytesimal'))
+sys.path.insert(0, os.path.abspath('../..'))
+# sys.path.insert(0, os.path.abspath('../../pytesimal'))
 sys.path.insert(0, os.path.abspath('../'))
