+++ pytesimal/draft_mantle_timestepping_2.py
@@ -8,16 +8,19 @@ Created on Thu Apr 8 16:24:52 2021.
 
 to do: need to move temp array set up outside of the discretisation function.
 """
-# import os
+
 import numpy as np
-# import pickle
-# import inspect
-# import sys
-import draft_core_functions_2
+
+# these relative imports are tripping sphinx up
+# want to change these anyway (so that mantle_timestepping function just
+# accepts these as arguments as opposed to the objects being
+# instantiated inside the mantle_timestepping function...
+
 import draft_mantle_properties
+import draft_core_functions_2
 
 
-class MantleBC: # TODO maybe instead include this in the setup function?
+class MantleBC:
     def __init__(self):
         pass
 
@@ -25,6 +28,32 @@ class MantleBC: # TODO maybe instead include this in the setup function?
         temperatures[-1, i] = temp_surface
         temperatures[0, i] = core_boundary_temperature
 
+
+class EnergyExtractedAcrossCMB:
+    """Calculate the energy extracted across the cmb in one timestep given the temperature of the mantle."""
+    def __init__(self, outer_r, timestep, radius_step):
+        self.radius = outer_r
+        self.dt = timestep
+        self.dr = radius_step
+        self.area = 4 * np.pi * self.radius**2
+
+    def __str__(self):
+        """Return string."""
+        return """Calculate the energy extracted across the core mantle boundary given the temperature of the mantle."""
+
+    def energy_extracted(self, mantle_temperatures, i, k):
+        """Calculate energy extracted in one timestep"""
+        energy = -self.area * k * (
+                (mantle_temperatures[0, i] - mantle_temperatures[1, i])
+                / self.dr) * self.dt
+        return energy
+
+    def power(self, mantle_temperatures, i, k):
+        """Calculate heat (power) extracted in one timestep"""
+        heat = -self.area * k * (
+                (mantle_temperatures[0, i] - mantle_temperatures[1, i])
+                / self.dr)
+        return heat
 
 def discretisation(
     latent,
@@ -89,7 +118,7 @@ def discretisation(
     else:
         dens = draft_mantle_properties.VariableDensity()
 
-    temp_list_mid_mantle = [temp_init]  # TODO - make this all possibly variable
+    temp_list_mid_mantle = [temp_init]
     temp_list_shal = [temp_init]
     temp_list_cmb_5 = [temp_init]
     temperatures[:, 0] = temp_init # this can be an array or a scalar
@@ -98,25 +127,25 @@ def discretisation(
     # instantiate boundary conditions
     boundary_conds = MantleBC()
     # instantiate core object and energy extracted
-    cmb_energy = draft_core_functions_2.EnergyExtractedAcrossCMB(r_core, timestep, dr)
+    cmb_energy = EnergyExtractedAcrossCMB(r_core, timestep, dr)
     core_values = draft_core_functions_2.IsothermalEutecticCore(temp=core_temp_init, melt=temp_core_melting,
                                                                 outer_r=r_core, inner_r=0, rho=core_density, cp=core_cp,
-                                                                core_latent_heat=core_latent_heat)
+                                                                core_latent_heat=core_latent_heat) # TODO move this out of the core
 
-    A_1list = [] # TODO make all these customisable
-    B_1list = []
-    C_1list = []
-    delt_list = []
-
-    A_1listcmb = []
-    B_1listcmb = []
-    C_1listcmb = []
-    delt_listcmb = []
-
-    A_1listshal = []
-    B_1listshal = []
-    C_1listshal = []
-    delt_listshal = []
+    # A_1list = []
+    # B_1list = []
+    # C_1list = []
+    # delt_list = []
+    #
+    # A_1listcmb = []
+    # B_1listcmb = []
+    # C_1listcmb = []
+    # delt_listcmb = []
+    #
+    # A_1listshal = []
+    # B_1listshal = []
+    # C_1listshal = []
+    # delt_listshal = []
 
     for i in range(1, len(times[1:]) + 1):
 
@@ -203,37 +232,37 @@ def discretisation(
 
                 temperatures[j, i] = temperatures[j, i - 1] + A_1 + B_1 + C_1
 
-            checkpoint_mm = radii[int(len(radii) / 2.0)]  # TODO just let checkpoints be fed in manually
-            checkpoint_cmb = radii[int(2.0 * len(radii) / 3.0)]
-            checkpoint_shal = radii[int(len(radii) / 3.0)]
-            if radii[j] == checkpoint_mm:
-                temp_list_mid_mantle.append(temperatures[j, i])
-                A_1list.append(A_1)
-                B_1list.append(B_1)
-                C_1list.append(C_1)
-                delt_list.append(
-                    ((temperatures[j, i - 1]) - (temperatures[j, i]))
-                )
-
-            elif radii[j] == checkpoint_cmb:
-                temp_list_cmb_5.append(temperatures[j, i])
-                A_1listcmb.append(A_1)
-                B_1listcmb.append(B_1)
-                C_1listcmb.append(C_1)
-                delt_listcmb.append(
-                    ((temperatures[j, i - 1]) - (temperatures[j, i]))
-                )
-
-            elif radii[j] == checkpoint_shal:
-                temp_list_shal.append(temperatures[j, i])
-                A_1listshal.append(A_1)
-                B_1listshal.append(B_1)
-                C_1listshal.append(C_1)
-                delt_listshal.append(
-                    ((temperatures[j, i - 1]) - (temperatures[j, i]))
-                )
-            else:
-                pass
+            # checkpoint_mm = radii[int(len(radii) / 2.0)]
+            # checkpoint_cmb = radii[int(2.0 * len(radii) / 3.0)]
+            # checkpoint_shal = radii[int(len(radii) / 3.0)]
+            # if radii[j] == checkpoint_mm:
+            #     temp_list_mid_mantle.append(temperatures[j, i])
+            #     A_1list.append(A_1)
+            #     B_1list.append(B_1)
+            #     C_1list.append(C_1)
+            #     delt_list.append(
+            #         ((temperatures[j, i - 1]) - (temperatures[j, i]))
+            #     )
+            #
+            # elif radii[j] == checkpoint_cmb:
+            #     temp_list_cmb_5.append(temperatures[j, i])
+            #     A_1listcmb.append(A_1)
+            #     B_1listcmb.append(B_1)
+            #     C_1listcmb.append(C_1)
+            #     delt_listcmb.append(
+            #         ((temperatures[j, i - 1]) - (temperatures[j, i]))
+            #     )
+            #
+            # elif radii[j] == checkpoint_shal:
+            #     temp_list_shal.append(temperatures[j, i])
+            #     A_1listshal.append(A_1)
+            #     B_1listshal.append(B_1)
+            #     C_1listshal.append(C_1)
+            #     delt_listshal.append(
+            #         ((temperatures[j, i - 1]) - (temperatures[j, i]))
+            #     )
+            # else:
+            #     pass
 
         # set top and bottom temperatures as fixed
         boundary_conds.dirichlet(temperatures,
@@ -252,19 +281,19 @@ def discretisation(
         temperatures,
         coretemp_array,
         latent,
-        temp_list_mid_mantle,
-        temp_list_shal,
-        temp_list_cmb_5,
-        A_1list,
-        B_1list,
-        C_1list,
-        delt_list,
-        A_1listcmb,
-        B_1listcmb,
-        C_1listcmb,
-        delt_listcmb,
-        A_1listshal,
-        B_1listshal,
-        C_1listshal,
-        delt_listshal,
+        # temp_list_mid_mantle,
+        # temp_list_shal,
+        # temp_list_cmb_5,
+        # A_1list,
+        # B_1list,
+        # C_1list,
+        # delt_list,
+        # A_1listcmb,
+        # B_1listcmb,
+        # C_1listcmb,
+        # delt_listcmb,
+        # A_1listshal,
+        # B_1listshal,
+        # C_1listshal,
+        # delt_listshal,
     )
