+++ tests/conftest.py
@@ -54,21 +54,21 @@ def temperature_timestepping():
     (mantle_temperature_array,
     core_temperature_array,
     latent,
-    temp_list_mid_mantle,
-    temp_list_shal,
-    temp_list_cmb_5,
-    A_1list,
-    B_1list,
-    C_1list,
-    delt_list,
-    A_1listcmb,
-    B_1listcmb,
-    C_1listcmb,
-    delt_listcmb,
-    A_1listshal,
-    B_1listshal,
-    C_1listshal,
-    delt_listshal,
+    # temp_list_mid_mantle,
+    # temp_list_shal,
+    # temp_list_cmb_5,
+    # A_1list,
+    # B_1list,
+    # C_1list,
+    # delt_list,
+    # A_1listcmb,
+    # B_1listcmb,
+    # C_1listcmb,
+    # delt_listcmb,
+    # A_1listshal,
+    # B_1listshal,
+    # C_1listshal,
+    # delt_listshal,
     )= mantle_timestepping.discretisation(
         latent,
         temp_init,
