+++ tests/test_newcoreandmantle_integration.py
@@ -22,21 +22,22 @@ def test_mtt_discretisation():
     (temperatures,
         coretemp,
         latent,
-        temp_list_mid_mantle,
-        temp_list_shal,
-        temp_list_cmb_5,
-        A_1list,
-        B_1list,
-        C_1list,
-        delt_list,
-        A_1listcmb,
-        B_1listcmb,
-        C_1listcmb,
-        delt_listcmb,
-        A_1listshal,
-        B_1listshal,
-        C_1listshal,
-        delt_listshal,) = mtt.discretisation(latent=[],
+        # temp_list_mid_mantle,
+        # temp_list_shal,
+        # temp_list_cmb_5,
+        # A_1list,
+        # B_1list,
+        # C_1list,
+        # delt_list,
+        # A_1listcmb,
+        # B_1listcmb,
+        # C_1listcmb,
+        # delt_listcmb,
+        # A_1listshal,
+        # B_1listshal,
+        # C_1listshal,
+        # delt_listshal,
+     ) = mtt.discretisation(latent=[],
                                              temp_init=1800.0,
                                              core_temp_init=1800,
                                              temp_core_melting=1200.0,
@@ -60,11 +61,11 @@ def test_mtt_discretisation():
                                              )
     temp_mean = np.mean(temperatures)
     coretemp_mean = np.mean(coretemp)
-    delt_listshal_mean = np.mean(delt_listshal)
+    # delt_listshal_mean = np.mean(delt_listshal)
 
     assert temp_mean == pytest.approx(1338.428508059057, 0.0000000001)
     assert coretemp_mean == pytest.approx(1800.0)
-    assert delt_listshal_mean == pytest.approx(0.0034806426277591207)
+    # assert delt_listshal_mean == pytest.approx(0.0034806426277591207)
     print("Success.")
 
 
@@ -78,21 +79,22 @@ def test_mtt_vardiscretisation():
     (temperatures,
         coretemp,
         latent,
-        temp_list_mid_mantle,
-        temp_list_shal,
-        temp_list_cmb_5,
-        A_1list,
-        B_1list,
-        C_1list,
-        delt_list,
-        A_1listcmb,
-        B_1listcmb,
-        C_1listcmb,
-        delt_listcmb,
-        A_1listshal,
-        B_1listshal,
-        C_1listshal,
-        delt_listshal,) = mtt.discretisation(latent=[],
+        # temp_list_mid_mantle,
+        # temp_list_shal,
+        # temp_list_cmb_5,
+        # A_1list,
+        # B_1list,
+        # C_1list,
+        # delt_list,
+        # A_1listcmb,
+        # B_1listcmb,
+        # C_1listcmb,
+        # delt_listcmb,
+        # A_1listshal,
+        # B_1listshal,
+        # C_1listshal,
+        # delt_listshal,
+     ) = mtt.discretisation(latent=[],
                                              temp_init=1800.0,
                                              core_temp_init=1800,
                                              temp_core_melting=1200.0,
@@ -116,10 +118,10 @@ def test_mtt_vardiscretisation():
                                              )
     temp_mean = np.mean(temperatures)
     coretemp_mean = np.mean(coretemp)
-    delt_listshal_mean = np.mean(delt_listshal)
+    # delt_listshal_mean = np.mean(delt_listshal)
     assert temp_mean == pytest.approx(1692.134212968852, 0.0000000001)
     assert coretemp_mean == pytest.approx(1800.0)
-    assert delt_listshal_mean == pytest.approx(1.012893130791781e-10)
+    # assert delt_listshal_mean == pytest.approx(1.012893130791781e-10)
     print("Success.")
 
 def test_mtt_discretisation_cold_mantle():
@@ -132,21 +134,22 @@ def test_mtt_discretisation_cold_mantle():
     (temperatures,
         coretemp,
         latent,
-        temp_list_mid_mantle,
-        temp_list_shal,
-        temp_list_cmb_5,
-        A_1list,
-        B_1list,
-        C_1list,
-        delt_list,
-        A_1listcmb,
-        B_1listcmb,
-        C_1listcmb,
-        delt_listcmb,
-        A_1listshal,
-        B_1listshal,
-        C_1listshal,
-        delt_listshal,) = mtt.discretisation(latent=[],
+        # temp_list_mid_mantle,
+        # temp_list_shal,
+        # temp_list_cmb_5,
+        # A_1list,
+        # B_1list,
+        # C_1list,
+        # delt_list,
+        # A_1listcmb,
+        # B_1listcmb,
+        # C_1listcmb,
+        # delt_listcmb,
+        # A_1listshal,
+        # B_1listshal,
+        # C_1listshal,
+        # delt_listshal,
+     ) = mtt.discretisation(latent=[],
                                              temp_init=1400.0,
                                              core_temp_init=1800,
                                              temp_core_melting=1200.0,
@@ -170,12 +173,12 @@ def test_mtt_discretisation_cold_mantle():
                                              )
     temp_mean = np.mean(temperatures)
     coretemp_mean = np.mean(coretemp)
-    delt_listshal_mean = np.mean(delt_listshal)
-    print(temp_mean, coretemp_mean, delt_listshal_mean)
+    # delt_listshal_mean = np.mean(delt_listshal)
+    # print(temp_mean, coretemp_mean, delt_listshal_mean)
 
     assert temp_mean == pytest.approx(1129.0036119517447, 0.0000000001)
     assert coretemp_mean == pytest.approx(1799.9984226292759)
-    assert delt_listshal_mean == pytest.approx(0.0012771823292144018)
+    # assert delt_listshal_mean == pytest.approx(0.0012771823292144018)
     print("Success.")
 
 
@@ -189,21 +192,22 @@ def test_mtt_vardiscretisation_cold_mantle():
     (temperatures,
         coretemp,
         latent,
-        temp_list_mid_mantle,
-        temp_list_shal,
-        temp_list_cmb_5,
-        A_1list,
-        B_1list,
-        C_1list,
-        delt_list,
-        A_1listcmb,
-        B_1listcmb,
-        C_1listcmb,
-        delt_listcmb,
-        A_1listshal,
-        B_1listshal,
-        C_1listshal,
-        delt_listshal,) = mtt.discretisation(latent=[],
+        # temp_list_mid_mantle,
+        # temp_list_shal,
+        # temp_list_cmb_5,
+        # A_1list,
+        # B_1list,
+        # C_1list,
+        # delt_list,
+        # A_1listcmb,
+        # B_1listcmb,
+        # C_1listcmb,
+        # delt_listcmb,
+        # A_1listshal,
+        # B_1listshal,
+        # C_1listshal,
+        # delt_listshal,
+     ) = mtt.discretisation(latent=[],
                                              temp_init=1200.0,
                                              core_temp_init=1800,
                                              temp_core_melting=1200.0,
@@ -227,10 +231,10 @@ def test_mtt_vardiscretisation_cold_mantle():
                                              )
     temp_mean = np.mean(temperatures)
     coretemp_mean = np.mean(coretemp)
-    delt_listshal_mean = np.mean(delt_listshal)
-    print(temp_mean, coretemp_mean, delt_listshal_mean)
+    # delt_listshal_mean = np.mean(delt_listshal)
+    # print(temp_mean, coretemp_mean, delt_listshal_mean)
 
     assert temp_mean == pytest.approx(1161.0663997556733, 0.0000000001)
     assert coretemp_mean == pytest.approx(1799.9985585325398)
-    assert delt_listshal_mean == pytest.approx(-7.581142508447556e-06)
+    # assert delt_listshal_mean == pytest.approx(-7.581142508447556e-06)
     print("Success.")
